Added a edit button
Added a edit button, this makes it so that a user cannot edit their profile before clicking on edit. When they click on edit they are able to make changes and will then be able to click on save changes once all the input validation has been passed, if not then the user will not be able to save their changes.